requirements.txt
model-bakery
Sphinx
Pillow
tox
coverage
django-debug-toolbar
